Added Application overhead. Added rounding to double values to correct loss of prcesion

+++ src/edu/uwo/csd/dcsim2/application/StaticWorkload.java
@@ -4,9 +4,9 @@ import edu.uwo.csd.dcsim2.core.Simulation;
 
 public class StaticWorkload extends Workload {
 
-	int workPerSecond;
+	double workPerSecond;
 	
-	public StaticWorkload(int workPerSecond) {
+	public StaticWorkload(double workPerSecond) {
 		super();
 		
 		this.workPerSecond = workPerSecond;
